# services/indexer/openapi.yaml
openapi: "3.1.0"
info:
  title: Linkora Indexer API
  version: "0.1.0"
  description: REST API for querying indexed on-chain data from the Linkora protocol.
servers:
  - url: http://localhost:3001/api
    description: Local development server
paths:
...
  /feed/following:
    get:
      summary: Get chronologically ordered posts from followed accounts
      operationId: getFollowingFeed
      tags: [Feed]
      parameters:
        - name: viewer
          in: query
          required: true
          schema:
            type: string
          description: Viewer Stellar address.
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/after"
      responses:
        "200":
          description: Paginated feed of posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedList"
        "400":
          $ref: "#/components/responses/BadRequest"
  /feed/explore:
    get:
      summary: Get hot-scored posts
      operationId: getExploreFeed
      tags: [Feed]
      parameters:
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/after"
      responses:
        "200":
          description: Paginated feed of posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedList"
        "400":
          $ref: "#/components/responses/BadRequest"
...
components:
  parameters:
...
    limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Maximum number of items to return.
...
    after:
      name: after
      in: query
      required: false
      schema:
        type: string
      description: Keyset cursor to fetch the next page.
  schemas:
...
    FeedPost:
      type: object
      required: [id, author, content, tip_total, like_count, created_at, cursor]
      properties:
        id:
          type: string
        author:
          type: string
        content:
          type: string
        tip_total:
          type: string
        like_count:
          type: string
        created_at:
          type: string
          format: date-time
        cursor:
          type: string
    FeedList:
      type: object
      required: [posts, limit, has_more]
      properties:
        posts:
          type: array
          items:
            $ref: "#/components/schemas/FeedPost"
        limit:
          type: integer
        has_more:
          type: boolean
        next_cursor:
          type: string
          nullable: true
...
    Error:
      type: object
      required: [error, code]
      properties:
        error:
          type: string
        code:
          type: string
  responses:
...
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: limit must be a positive integer
            code: INVALID_QUERY